# RFC: Code refactor