--- a/REQUIRE
+++ b/REQUIRE
@@ -7,3 +7,4 @@ StaticArrays
 Colors 0.7.0
 ColorVectorSpace 0.2
 FixedPointNumbers 0.3
+Compat 0.17