I am in the process of cleaning up the code a bit and writing test. This is not even close to merge ready, but I already made some changes that may be controversial so I just wanted to give the option of early feedback. 

One thing I did is move related code blocks into files, which granted may at this point not really be necessary. Yet it seems better to do it early, while the code-base is small enough and changes are easier to follow