@Evizero 2017-03-16
I am playing around with updating restrict. It seems to work, but as of now my implementation is a factor 10 slower. :(

old

```
julia> @benchmark restrict($img)
BenchmarkTools.Trial: 
  memory estimate:  772.72 KiB
  allocs estimate:  32
  --------------
  minimum time:     761.340 μs (0.00% GC)
  median time:      813.791 μs (0.00% GC)
  mean time:        823.768 μs (1.89% GC)
  maximum time:     1.334 ms (34.95% GC)
  --------------
  samples:          6023
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
```

new

```
julia> @benchmark restrict($img)
BenchmarkTools.Trial: 
  memory estimate:  772.97 KiB
  allocs estimate:  48
  --------------
  minimum time:     8.575 ms (0.00% GC)
  median time:      8.839 ms (0.00% GC)
  mean time:        8.868 ms (0.17% GC)
  maximum time:     9.572 ms (4.68% GC)
  --------------
  samples:          564
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
```