@Evizero 2017-04-10
I think I addressed everything. ready for another round of review and/or merge