@timholy 2017-04-10
Thanks. I don't need to see this get through CI again, so I say merge. I'll let you decide whether this should be squashed in one big commit or whether smaller commits are more appropriate.