@Evizero 2017-04-10
I think a squash merge is the only sensible way to go for changes like this one, since it is such diff-breaking code move